rename cluster new cluster node file
